+++ library/Rules/Countable.php
@@ -11,12 +11,10 @@
 
 namespace Respect\Validation\Rules;
 
-class ArrayVal extends AbstractRule
+class Countable extends AbstractRule
 {
     public function validate($input)
     {
-        return is_array($input) || ($input instanceof \ArrayAccess
-            && $input instanceof \Traversable
-            && $input instanceof \Countable);
+        return is_array($input) || $input instanceof \Countable;
     }
 }